requirements.txt
opencv-python>=4.10.0
python-statemachine>=2.5.0